전체적인 백엔드 작업완료

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   try {
@@ -11,7 +12,8 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner"); //populate가 ref 참조해서 해당하는 schema의 속성값 갖다줌
+  console.log(video);
   if (video) {
     return res.render("watch", { pageTitle: video.title, video });
   }
@@ -20,9 +22,16 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  console.log(video.owner, "그리고", _id);
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
@@ -45,17 +54,26 @@ export const getUpload = (req, res) => {
   return res.render("upload", { pageTitle: "Upload video!" });
 };
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file; // req.file.path의 값을 fileUrl이라는 변수에 초기화 한다는 뜻
   const { title, description, hashtags } = req.body;
   try {
-    const video = Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
       hashtags: Video.formatHashtags(hashtags),
+      owner: _id,
       meta: {
         views: 0,
         rating: 0,
       },
     });
+    const ownUser = await User.findById(_id);
+    ownUser.videos.push(newVideo._id);
+    ownUser.save();
     // await video.save();
     return res.redirect("/");
   } catch (error) {
@@ -68,8 +86,21 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
-  await Video.findByIdAndDelete(id);
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    //비디오가 없는 경우
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    //다른 이용자가 비디오를 삭제하려는 경우
+    console.log("어림도 없다 암");
+    return res.status(403).redirect("/");
+  }
   // delete video
+  await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
